refactor: setter 제거
